% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_shortcuts.R
\name{editor_shortcuts}
\alias{editor_shortcuts}
\title{Various snippets}
\description{
RStudio add-in to insert various snippets: \itemize{
   \item \code{curly_braces()} - insert curly braces;
}.


These functions internally use the following functions:

\code{rs_enclose_all_with_lines()} -
that adds lines above and below the selection.

\code{rs_enclose_selection_with()} -
that encloses selection with some text. If there is
no selection, it encloses the cursor.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
This function is by Gagznev https://github.com/GegznaV/addin.tools
}
\concept{Editor add-ins}
