% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kispaddins-package.R
\docType{package}
\name{kispaddins-package}
\alias{kispaddins}
\alias{kispaddins-package}
\title{kispaddins: KİSP addin functions}
\description{
After installation, the package provides Rstudio add-ins. These can be used by the dropdown "Addins" menuor can be assigned keyboard shortcuts.
}
\keyword{internal}
