% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{rs_get_context}
\alias{rs_get_context}
\title{Retrieve information about an active RStudio document.
This function is by Gagznev https://github.com/GegznaV/addin.tools
Wrapper for \code{\link[rstudioapi]{getActiveDocumentContext}()}.}
\usage{
rs_get_context()
}
\description{
Retrieve information about an active RStudio document.
This function is by Gagznev https://github.com/GegznaV/addin.tools
Wrapper for \code{\link[rstudioapi]{getActiveDocumentContext}()}.
}
